added response model to api

+++ app/data_models/predictions.py
@@ -6,4 +6,4 @@ class ProductTierPrediction(BaseModel):
 
 
 class DetailViewsPrediction(BaseModel):
-    detail_views: int
+    detail_views: float
